Fix : 자료형 수정

+++ src/main/java/backend/mossy/boundedContext/payout/app/PayoutFacade.java
@@ -3,6 +3,7 @@ package backend.mossy.boundedContext.payout.app;
 import backend.mossy.boundedContext.payout.domain.Payout;
 import backend.mossy.boundedContext.payout.domain.PayoutCandidateItem;
 import backend.mossy.global.rsData.RsData;
+import backend.mossy.shared.market.dto.event.OrderDto;
 import backend.mossy.shared.payout.dto.event.SellerDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -27,7 +28,7 @@ public class PayoutFacade {
     }
 
     @Transactional
-    public Payout createPayout(int payeeId) {
+    public Payout createPayout(Long payeeId) {
         return payoutCreatePayoutUseCase.createPayout(payeeId);
     }
